学校安装设备数量管理

+++ src/main/java/com/dmkj/ljadmin/common/security/LoginUserDetails.java
@@ -16,7 +16,7 @@ public class LoginUserDetails implements UserDetails {
 
     private UserDto user;
 
-    private List<Long> dataScopes;
+    private List<Integer> dataScopes;
 
     private List<AuthorityDto> authorities;
